Got the theory for building packages. We build a graph with edges for depends and required by's going down a depends edges is considered going in the down direction going down a required by edges is considered going in the up direction if we go all the way down, we will find a base package, and we can build that safely after each one is buit, we re-trace the graph going as far down as possible until we hit bottom and build that one a circular dependancy can be detected by keeping a list of visited packages.  if we find a package already on our list while going down, we assume we hit bottom and we build all packages between the matched indice and the one at the bottom once, and then build the first package again.  an example of this is the linux buildchain described as so: binutilsi->gcc->glibc->binutils->gcc in this case, we have nested circular dependancies. the first case occurs in the binutils->...->bintuils case and the second in the gcc->...->gcc case.  in this case, glibc would be built twice and gcc would be built three times by this algorithm.
This should ensure a proper package output.  Also some packages may need
a first pass version like libav, which requires --enable-libx264 to be
turned off on the first pass.  I will likely use a mangle chain in the
actual builder to handle this.  Also in the mangle chain may be a way of
modifying the build list from the mangle chain.  this way, if there is
an incompatible API change in the required by packages, we can avoid
building this package. though this is a project for the future.